refactor: implementing new refresh session page
