+++ src/app/(smal)/auth/refresh/page.client.tsx
@@ -4,19 +4,8 @@ import { trpc } from "@/app/_trpc/client";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect } from "react";
 import { clearAuthData } from "@/utils/authStorage";
-import { cn } from "@/lib/utils";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle
-} from "@/components/ui/card";
 
-const SessionRefresher = ({
-  className,
-  ...props
-}: React.ComponentProps<"div">) => {
+const SessionRefresher = () => {
   const router = useRouter();
   const searchParams = useSearchParams();
   const redirect = searchParams.get("redirect") || "/";
@@ -46,21 +35,7 @@ const SessionRefresher = ({
     refreshSession({ refreshToken });
   }, [redirect]);
 
-  return (
-    <div className={cn("flex flex-col gap-6", className)} {...props}>
-      <Card className="border-0 shadow-none">
-        <CardHeader>
-          <CardTitle>Session Refresh</CardTitle>
-          <CardDescription>
-            We are refreshing your session. Please wait...
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="text-center">
-          <p>Refreshing your session...</p>
-        </CardContent>
-      </Card>
-    </div>
-  );
+  return null;
 };
 
 export default SessionRefresher;
