--- src/app/auth/refresh/page.tsx
@@ -1,16 +1,0 @@
-import { Suspense } from "react";
-import SessionRefresher from "./page.client";
-
-const Page = () => {
-  return (
-    <div className="flex justify-center w-full">
-      <div className="w-full max-w-sm">
-        <Suspense>
-          <SessionRefresher />
-        </Suspense>
-      </div>
-    </div>
-  );
-};
-
-export default Page;
